Developing new Pole Figure generation code with Experimental filter and tool.

+++ Source/DREAM3DLib/Common/DREAM3DMath.h
@@ -129,7 +129,7 @@ namespace DREAM3D
   namespace Constants
   {
     static const float k_Pi = static_cast<float>(M_PI);
-    static const float k_SqrtPi = sqrt(M_PI);
+    static const double k_SqrtPi = sqrt(M_PI);
     static const double k_2Pi = 2.0 * M_PI;
     static const double k_1OverPi = 1.0/M_PI;
     static const double k_PiOver180 = M_PI/180.0;
